fix(ci): python-coverage-shards — split test_worker_part_038 under the 1500-line gate
test_worker_part_038.py grew to 1516 lines (the epoch-read fencing tests
added during PR #421 work pushed it over MAX_FIRST_PARTY_FILE_LINES=1500),
failing test_core_decomposition_maintainability.

Move the eight tail tests (DB-closed event handling, dispatch-limit
helpers, active-salvage bookkeeping bounds, and the _safely_* failure
isolation/cancel-propagation paths) into a new self-contained
test_worker_part_047.py carrying only the imports, fixtures, and helpers
those tests use. Drop the now-unused imports left behind in part_038.

No behavior change; all 20 tests are preserved (12 + 8) and pass.
